Added plotly subpackage to renderers package for PlotlyRenderer obj.
Plotly subpackage includes the plotly_renderer module which contains the
PlotlyRenderer object which inherits from the renderers.base.Renderer
class. It reimplements the suggested methods commented on in
renderers.base.Renderer.

The plotly_utils module is included to separate basic style and
translation functionality from the main objective of the plotly_renderer
module which is to construct the data and layout dictionaries.

The PlotlyRenderer may differ slightly from the functionality of other
renderers as it just 'updates' its member dictionaries (self.data and
self.layout) as the Exporter object crawls around the matplotlib figure.
When the Exporter reaches the end of its crawl and the renderer 'closes'
the figure, PlotlyRenderer makes a call to plotly with the data and
layout dictionaries. This causes your browser to open a page pointing to
your new plotly plot!

It may or may not be of use to keep the renderers separated by
subpackages as was implemented with the plotly subpackage. TBD.