Added test_plotly module to tests package to use with PlotlyRenderer.
The test_plotly function includes a public username and api_key (feel
free to put your own in in the test_plotly.py file). It presently
defines a function, plot_sin, which will create a matplotlib figure,
instantiate a PlotlyRenderer, instantiate an Exporter with the
PlotlyRenderer, and use the Exporter's 'run' function.

The output is mimicked from the test_basic module. This is helpful to
understand the interplay between the Exporter and PlotlyRenderer
objects. A simple output assertion should not throw any exceptions.

A plot will show up in your browser if this function is called.