Changes Exporter._crawl_fig and _crawl_ax methods. (Experimental!)
The matplotlib package has a getp function that can be applied to
figures, axes, and lines. The old methods of populating the properties'
dictionaries are replaced with dictionary comprehensions which use
lists, which could be an immutable iterable, from the utils module.

For compatibility with the previous commits, certain keywords were added
back in after the dictionary comprehensions were built.