Better test function and nicer looking example for subplots.
The test function in the 'tests' dir now uses a custom function for
comparing two dictionaries which outputs the exact dictionary entry upon
a comparison failure. This is used as a message for the assertions so
that errors can be found quickly.

Example for subplots now uses matplotlib.gridspec and includes extra
'wspace' and 'hspace' so that the resulting plot ticks can be clearly
seen. It shows how complicated subplot structures can be ported to
Plotly.