Added plotly_utils.clean_dict(). Removes None. 'None', 'none', and {}.
JSON dictionaries sent to plotly should not have entries that contain
'none', 'None', None, or are themselves empty dictionaries ({}). This
should be updated at some point to disregard entries that are expected
to be labels so that users could use 'None' or 'none' to label a plot.
This is pretty unlikely, but should be changed.

clean_dict(node, partent=None, node_key=None) is a recursive function to
clean out a dictionary. Only nodes that are dictionaries are looped back
through the function. The JSON structures are dictionary-based, while
matplotlib structures are lists, no unintentional deletion problems have
surface yet.