Added user_friendly dictionary repair functions to plotly_utils.
There are many situations where and 'x1', 'y1', 'xaxis1', etc. show up
in the final collections for layout or data.

To make future fixes simpler the following system was put in place:

- To rename a key add the following to the *_key_repairs dict:
(key_path_tup): new_key

- To repair a value add the following to the *_val_repairs dict:
(key_path_tup): {bug: fix}

'key_path_tup'** is a tuple of dictionary keys. 'new_key' is the key you
would like to rename 'key_path_tuple[-1]' with. 'bug' is the value that
needs to be changed to 'fix'.

** key_path_tup will work ACROSS lists that are nested within plotly
JSON structures. For example, 'annotations' is a key in layout, but
layout['annotations'] is a list of dictionaries. These new functions
handle this inconsistency by splitting into a for loop each time this
happens and continuing along the remaining key_path_tup for each
dictionary in the list.

PlotlyRenderer class was updated to use this new functionality.