Workaround for annotation transforms and axis label size.
These are hacky workarounds:

- Transform for annotation gives the transformation necessary to get
annotation coordinates into display coordinates. However, the
transformation to put these coordinates into data coordinates is not
given. A solution might be to either pass the axes object upon calling
draw_text. The 'private' _current_ax object from the base renderer class
was used to transform annotations to data coordinates (what plotly
generally uses)

- Axis labels are children of xaxis and yaxis instances, which are
children of the current axes object. When the label is passed to the
renderer, it would be helpful to either get this specific text instance
or the size. I used the current axis to get at the xaxis/yaxis objects
and then filted their lists of children until i found the label that
matched the one mplexporter passes to the renderer. I don't know that
there is a way to get around having to sort through the children and
compare strings. This probably won't break down for 'normal' use, but a
pathological example could make it do weird things.