Margins are no longer just set to zero. Domains adjusted.
The bbox for mpl domains is in figure coordinates. Plotly's plotting
area should have a domain of [0,1] for x and y directions. The bbox for
incoming mpl plots are scaled by the maximum and minimum locations of
any of the subplots in the figure. That is, mpl_x_min is mapped to 0,
mpl_x_max is mapped to 1, mpl_y_min is mapped to 0, and mpl_y_max is
mapped to 1.