Added 'fontsize' to axis properties.
Added a few lines to grab the tick font size. You might want to change
the name of this dictionary key to be something else. I left it as
'fontsize' because that's what mpl calls it! The real confusion is that
it's in the axis dict, but should not be confused with the axis label
size. You might consider changing it to 'ticksize' or something later
on?

The lines simply make a list of the font sizes for the ticks and checks
to see that they all match. I'm sure it's not 100% bug-free, i.e., if
the font size entires aren't numbers (None, '', 'none', etc.). However,
if they're not all the same it returns a None value.

As a side note, I really like being able to assume an exporter value is
either a valid entry or None. I think this format should be continued.