Added horizontal and vertical alignment to non-'data' text objects.
Plotly now allows users to specify the anchor for each annotation in a
plot. This means that it can use the horizontal and vertical alignments
stored in mpl figures to better re-create mpl plots.

A simple conversion function translates mpl alignment words to plotly
alignment words (most of them are the same). See
plotly_utils.convert_alignment.

Currently, text objects that are anchored to x,y data are allowed to be
'auto' anchored in plotly. This is to allow annotation arrows to show up
appropriately when using the site after a fig_to_plotly call. This will
be changed soon.