Removed `lines` and `markers` variables from draw_line in exporter.
If there is no `dasharray` for a Line2D obj, the `linestyle` dict is set
to `None`. Same goes when `marker` is not populated.

This way, the `base` renderer uses `linestyle` and `markerstyle` in
`draw_marked_line` to decide which or both of the `draw_line` and
`draw_markers` methods to call.