Make FuncFormatter work without FixedLocator.
It is impossible for exported plot to ever call the python function, so
we need to decide what to do when the user uses a FuncFormatter.

I decide that it's most meaningful to convert it to "fixed" ticks at
export-time, which is what this code does. However, it still keeps the
type as "func" so the consumer of export may choose what to do.

mpld3 just treats "func" and "fixed" identically. We could argue about
changing to "fixed" here at export-time; I think keeping "func" conveys
more information, but I don't feel strongly about it.

The current state for FuncFormatter is that it does convert the labels
at export-time only if locator is fixed. I argue this is strictly worse.

At the same time, I also made the unit-test assert util print much more
human interpretable test failure outputs, see comment.